+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 1.21.0
 
 Enabled multiple drafts functionality.
+Added favorite site functionality.
+Added moth trap weather attributes.
+Added the Slovak language.
 Added Saint Helena country.
 Added specific stage attributes for Dragonfly species.
 Added Dragonfly common names.